Add depth checks for cyclical / nested types
This change affects all unmarshaling for nested values but prevents
infinitely recursing while trying to unmarshal cyclical types.

This adds a max depth of 8 for structs where no fields are defined
whatsoever. The counter is reset if a value if the struct containing
the field has any other field set. If the struct has no other fields,
then depth cannot be reset right now.

This is done by refactoring Reader into a separate readState type that
tracks the current depth and, possibly later, other information about
read state. The readState, for the most part, is just a Reader wrapper.

Further improvements could be made to use an optional interface to scan
for known keys. This just deals with the immediate problem of unbounded
recursion.

Fixes #4.